requirements.txt
# Core framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlmodel==0.0.14
sqlalchemy==2.0.25
alembic==1.13.1

# Scheduling
apscheduler==3.10.4

# HTTP & RSS
httpx==0.26.0
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.1.0
readability-lxml==0.8.1

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# LLM clients
anthropic==0.18.1
openai==1.10.0

# Email
sendgrid==6.11.0
python-dotenv==1.0.1

# WhatsApp/Telegram (optional)
twilio==8.12.0
python-telegram-bot==20.8

# Templates
jinja2==3.1.3
markdown==3.5.2

# Utilities
python-dateutil==2.8.2
pytz==2024.1
pyyaml==6.0.1
hashlib-additional==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx-mock==0.1.2
